+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/datagen/model/ModBlockStateProvider.kt
@@ -226,6 +226,7 @@ class ModBlockStateProvider(
 				textureOther,
 				textureOther
 			)
+			.texture("particle", textureTop)
 
 		simpleBlockWithItem(block, model)
 	}