+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added the Forest and Flower Forest to the biome tag `#irregular_implements:nature_core_oak`
 - The Notification Interface now defaults to having no text in the title and description
+- The Player Interface no longer renders the owner's skull above itself if there's a block there
 
 ### Fixed