+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Notification Interface not having player inventory slots (#22)
 - Notification Interface wasn't saving the icon item
 - Couldn't type the letter `E` in the Notification Interface text boxes (#22)
+- Broken particle when mining the Player Interface (#24)
 
 # 1.3.1